applied command pattern and rabbit mq on messages flow
