+++ traderie/messages-service/src/main/java/com/massivelyflammableapps/messages/commands/chats/ChangeAcceptStatusCommand.java
@@ -1,6 +1,5 @@
-package com.massivelyflammableapps.messages.commands;
+package com.massivelyflammableapps.messages.commands.chats;
 
-import java.util.List;
 import java.util.UUID;
 
 import com.massivelyflammableapps.messages.model.Chat;
@@ -16,13 +15,13 @@ import lombok.RequiredArgsConstructor;
 @AllArgsConstructor
 @RequiredArgsConstructor
 @NoArgsConstructor
-public class GetUserChatsCommand extends ChatsAbstractCommand {
+public class ChangeAcceptStatusCommand extends ChatsAbstractCommand {
     private ChatService chatService;
 
     @NonNull
-    private UUID userId;
+    private UUID chatId;
 
-    public List<Chat> execute() {
-        return chatService.getUserChats(userId);
+    public Chat execute() {
+        return chatService.changeAcceptStatus(chatId);
     }
 }
