+++ traderie/messages-service/src/main/java/com/massivelyflammableapps/messages/commands/chats/ChangeArchiveStatusCommand.java
@@ -1,4 +1,4 @@
-package com.massivelyflammableapps.messages.commands;
+package com.massivelyflammableapps.messages.commands.chats;
 
 import java.util.UUID;
