+++ traderie/messages-service/src/main/java/com/massivelyflammableapps/messages/commands/chats/ChatsInvoker.java
@@ -1,4 +1,4 @@
-package com.massivelyflammableapps.messages.commands;
+package com.massivelyflammableapps.messages.commands.chats;
 
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.beans.factory.annotation.Autowired;
