+++ traderie/messages-service/src/main/java/com/massivelyflammableapps/messages/controller/MessagesController.java
@@ -27,7 +27,7 @@ public class MessagesController {
     }
 
     @PostMapping
-    public ResponseEntity<Message> postMessage(@RequestBody Message request) {
+    public Message postMessage(@RequestBody Message request) {
         return messageService.postMessage(request);
     }
